Add licenses
The config for awesome is licensed under GPLv2, because awesome uses
that license, and my config is based on the default config.

Everything else is licensed under the MIT License. This is allowed since
it has nothing to do with the awesome config, even though it is in the
same repository.