.gitconfig: Remove user.email
I use different email addresses for work and private projects, but
commit on both from the same computers. When cloning a new repo it is
easy to forget to set the correct email address if it is another than
the default.

By not having a default email address set, git will complain if I have
not set it in the repo. This way, I don't have to remember it myself.